♻️ Refactor: 마이페이지 index 분리, 모달 index에서 제외, 호출관련 이슈 발생 정리(#126)
