+++ src/components/mypage/myinfo/PasswordChange.tsx
@@ -7,9 +7,9 @@ import {
   ModalFooter,
   ModalTitle,
   ModalMain,
-  Button,
-} from '@/components'
-import { MypageInputField } from './MypageInputField'
+} from '@/components/common/Modal'
+import { Button } from '@/components'
+import { MypageInputField } from '@/components/mypage/myinfo/MypageInputField'
 
 export const PasswordChange = () => {
   return (
