+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import LandingPage from '@/pages/LandingPage'
-import { MyInfo } from '@/pages/MyInfo'
+import { MyInfo } from '@/pages/my-page/MyInfo'
 import BookmarkedRecruitment from '@/pages/my-page/BookmarkedRecruitment'
 import Login from '@/pages/Login'
