+++ src/pages/my-page/BookmarkedRecruitment.tsx
@@ -1,5 +1,6 @@
-import { ListItemSkeleton, BookmarkedRecruitmentCard } from '@/components'
+import { ListItemSkeleton } from '@/components'
 import { Input, InputIcon } from '@/components/common/input'
+import { BookmarkedRecruitmentCard } from '@/components/mypage'
 import EmptyDataState from '@/components/common/State/EmptyDataState'
 import { useBookmarkedRecruitment } from '@/hooks/api'
 import { useDebounce } from '@/hooks/useDebounce'
